Remove ruamel.yaml from all-test dependencies
ruamel isn't building on python 3.9 or 3.10. Upgrading ruamel may fix
that, but the yaml tests don't seem to be passing anyway, and weren't
included in CI. So rather than try to dig in right now to fix them, just
push the dependency fix to whoever decides to revive the yaml tests.